next commit

+++ oghma_core/libmemory/light_zxyl_long_double.c
@@ -1,29 +1,27 @@
 //
-// General-purpose Photovoltaic Device Model gpvdm.com - a drift diffusion
-// base/Shockley-Read-Hall model for 1st, 2nd and 3rd generation solarcells.
-// The model can simulate OLEDs, Perovskite cells, and OFETs.
+// OghmaNano - Organic and hybrid Material Nano Simulation tool
+// Copyright (C) 2008-2022 Roderick C. I. MacKenzie r.c.i.mackenzie at googlemail.com
 //
-// Copyright 2008-2022 Roderick C. I. MacKenzie https://www.gpvdm.com
-// r.c.i.mackenzie at googlemail.com
-//
+// https://www.oghma-nano.com
+// 
 // Permission is hereby granted, free of charge, to any person obtaining a
 // copy of this software and associated documentation files (the "Software"),
 // to deal in the Software without restriction, including without limitation
-// the rights to use, copy, modify, merge, publish, distribute, sublicense,
+// the rights to use, copy, modify, merge, publish, distribute, sublicense, 
 // and/or sell copies of the Software, and to permit persons to whom the
 // Software is furnished to do so, subject to the following conditions:
-//
+// 
 // The above copyright notice and this permission notice shall be included
 // in all copies or substantial portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
 // OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
 // THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE 
 // SOFTWARE.
-//
+// 
 
 /** @file memory_basic.c
 @brief memory functions for 3D arrays
@@ -42,26 +40,26 @@
 #include "memory.h"
 
 
-void malloc_light_zxyl_long_double(struct dimensions *dim, long double * (****var))
+void malloc_light_zxyl_long_double(struct dimensions *dim, gdouble * (****var))
 {
-	malloc_4d( (void *****)var, dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(long double));
+	malloc_4d( (void *****)var, dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(gdouble));
 }
 
 
 
-void free_light_zxyl_long_double(struct dimensions *dim, long double * (****in_var))
+void free_light_zxyl_long_double(struct dimensions *dim, gdouble * (****in_var))
 {
-	free_4d( (void *****)in_var, dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(long double));
+	free_4d( (void *****)in_var, dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(gdouble));
 }
 
-void cpy_light_zxyl_long_double(struct dimensions *dim, long double * (****out), long double * (****in))
+void cpy_light_zxyl_long_double(struct dimensions *dim, gdouble * (****out), gdouble * (****in))
 {
-	free_4d( (void *****)out, dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(long double));
-	malloc_4d( (void *****)out, dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(long double));
-	cpy_4d( (void *****)out, (void *****)in,dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(long double));
+	free_4d( (void *****)out, dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(gdouble));
+	malloc_4d( (void *****)out, dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(gdouble));
+	cpy_4d( (void *****)out, (void *****)in,dim->zlen, dim->xlen, dim->ylen,dim->llen,sizeof(gdouble));
 }
 
-void flip_light_zxyl_long_double_y(struct simulation *sim, struct dimensions *dim,long double **** data)
+void flip_light_zxyl_long_double_y(struct simulation *sim, struct dimensions *dim,gdouble **** data)
 {
 	int x=0;
 	int y=0;
@@ -69,7 +67,7 @@ void flip_light_zxyl_long_double_y(struct simulation *sim, struct dimensions *di
 	int l=0;
 
 
-	long double ****temp=NULL;
+	gdouble ****temp=NULL;
 
 	malloc_light_zxyl_long_double(dim, &temp);
 
@@ -110,7 +108,7 @@ void flip_light_zxyl_long_double_y(struct simulation *sim, struct dimensions *di
 	free_light_zxyl_long_double(dim, &temp);
 }
 
-void div_light_zxyl_long_double(struct dimensions *dim, long double ****data,long double val)
+void div_light_zxyl_long_double(struct dimensions *dim, gdouble ****data,gdouble val)
 {
 	int x=0;
 	int y=0;
@@ -133,7 +131,7 @@ void div_light_zxyl_long_double(struct dimensions *dim, long double ****data,lon
 
 }
 
-void memset_light_zxyl_long_double(struct dimensions *dim, long double ****data,int val)
+void memset_light_zxyl_long_double(struct dimensions *dim, gdouble ****data,int val)
 {
 	int x=0;
 	int y=0;
@@ -145,14 +143,14 @@ void memset_light_zxyl_long_double(struct dimensions *dim, long double ****data,
 		{
 			for (y = 0; y < dim->ylen; y++)
 			{
-				memset(data[z][x][y], val, dim->llen * sizeof(long double ));
+				memset(data[z][x][y], val, dim->llen * sizeof(gdouble ));
 			}
 		}
 	}
 
 }
 
-void memset_light_zxyl_long_double_y(struct dimensions *dim, long double ****data,int z, int x, int l,long double val)
+void memset_light_zxyl_long_double_y(struct dimensions *dim, gdouble ****data,int z, int x, int l,gdouble val)
 {
 	int y=0;
 	for (y = 0; y < dim->ylen; y++)
